+++ tests/tests_storage/test_save_new_available_bikes.py
@@ -1,0 +1,26 @@
+from app.models import Bike
+from app.storage import get_available_bike_list, save_new_available_bikes
+
+
+async def test_save_new_available_bikes_happy_path(fixture_empty_available_bike_list):
+    available_bikes_list = [
+        Bike(
+            id='spectral_125_cf_9',
+            title='Spectral 125 CF 9',
+            link='https://www.canyon.com/en-cz/mountain-bikes/trail-bikes/spectral-125/cf/spectral-125-cf-9/3179.html?dwvar_3179_pv_rahmenfarbe=SR',
+            family='Spectral',
+            model='125 CF 9',
+        ),
+        Bike(
+            id='exceed_cf_7',
+            title='Exceed CF 7',
+            link='https://www.canyon.com/en-cz/mountain-bikes/cross-country-bikes/exceed/cf/exceed-cf-7/3128.html?dwvar_3128_pv_rahmenfarbe=WH%2FMC',
+            family='Exceed',
+            model='CF 7',
+        )
+    ]
+
+    res = await save_new_available_bikes(available_bikes_list)
+
+    assert res is None
+    assert await get_available_bike_list() == available_bikes_list
